docker/mqtt/requirements.txt
paho-mqtt
pytcs-tecnoalarm